Only use a single, correct ResponseBuilder -- fixes headers
