+++ azure-function-http/src/main/java/io/micronaut/azure/function/http/AzureFunctionHttpResponse.java
@@ -15,15 +15,13 @@
  */
 package io.micronaut.azure.function.http;
 
+import com.microsoft.azure.functions.HttpRequestMessage;
 import com.microsoft.azure.functions.HttpResponseMessage;
-import com.microsoft.azure.functions.HttpStatusType;
 import io.micronaut.core.annotation.Internal;
 import io.micronaut.core.annotation.Nullable;
-import io.micronaut.core.convert.ArgumentConversionContext;
 import io.micronaut.core.convert.ConversionService;
 import io.micronaut.core.convert.value.MutableConvertibleValues;
 import io.micronaut.core.convert.value.MutableConvertibleValuesMap;
-import io.micronaut.core.util.CollectionUtils;
 import io.micronaut.http.CaseInsensitiveMutableHttpHeaders;
 import io.micronaut.http.HttpHeaders;
 import io.micronaut.http.HttpStatus;
@@ -40,14 +38,7 @@ import java.io.BufferedWriter;
 import java.io.ByteArrayOutputStream;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 
 /**
  * Azure implementation of {@link ServletHttpResponse}.
@@ -63,6 +54,7 @@ public final class AzureFunctionHttpResponse<B> implements ServletHttpResponse<H
 
     private final ByteArrayOutputStream body = new ByteArrayOutputStream();
     private int status = HttpStatus.OK.getCode();
+    private final HttpRequestMessage<Optional<String>> azureRequest;
     private  final MutableHttpHeaders headers;
     private final BinaryContentConfiguration binaryContentConfiguration;
     private MutableConvertibleValues<Object> attributes;
@@ -70,9 +62,11 @@ public final class AzureFunctionHttpResponse<B> implements ServletHttpResponse<H
     private String reason = HttpStatus.OK.getReason();
 
     public AzureFunctionHttpResponse(
+        HttpRequestMessage<Optional<String>> azureRequest,
         ConversionService conversionService,
         BinaryContentConfiguration binaryContentConfiguration
     ) {
+        this.azureRequest = azureRequest;
         this.headers = new CaseInsensitiveMutableHttpHeaders(conversionService);
         this.binaryContentConfiguration = binaryContentConfiguration;
     }
@@ -157,7 +151,9 @@ public final class AzureFunctionHttpResponse<B> implements ServletHttpResponse<H
             LOG.trace("Using Azure Function HTTP Response Builder");
             return builder.build();
         }
-        HttpResponseMessage.Builder responseBuilder = new ResponseBuilder().status(com.microsoft.azure.functions.HttpStatus.valueOf(status));
+        HttpResponseMessage.Builder responseBuilder = azureRequest.createResponseBuilder(
+            com.microsoft.azure.functions.HttpStatus.valueOf(status)
+        );
         getHeaders().forEach((s, strings) -> {
             for (String string : strings) {
                 responseBuilder.header(s, string);
@@ -171,83 +167,5 @@ public final class AzureFunctionHttpResponse<B> implements ServletHttpResponse<H
             }
         }
         return responseBuilder.build();
-    }
-
-    private static class ResponseBuilder implements HttpResponseMessage.Builder, HttpResponseMessage, HttpHeaders {
-        private HttpStatusType status = com.microsoft.azure.functions.HttpStatus.OK;
-        private final Map<String, List<String>> headers = new LinkedHashMap<>(3);
-        private Object body;
-
-        @Override
-        public Builder status(HttpStatusType status) {
-            this.status = status;
-            return this;
-        }
-
-        @Override
-        public Builder header(String key, String value) {
-            headers.computeIfAbsent(key, k -> new ArrayList<>()).add(value);
-            return this;
-        }
-
-        @Override
-        public Builder body(Object body) {
-            this.body = body;
-            return this;
-        }
-
-        @Override
-        public HttpResponseMessage build() {
-            return this;
-        }
-
-        @Override
-        public HttpStatusType getStatus() {
-            return status;
-        }
-
-        @Override
-        public String getHeader(String key) {
-            List<String> v = headers.get(key);
-            if (CollectionUtils.isNotEmpty(v)) {
-                return v.iterator().next();
-            }
-            return null;
-        }
-
-        @Override
-        public Object getBody() {
-            return this.body;
-        }
-
-        @Override
-        public List<String> getAll(CharSequence name) {
-            List<String> values = headers.get(name.toString());
-            if (values != null) {
-                return Collections.unmodifiableList(values);
-            }
-            return Collections.emptyList();
-        }
-
-        @Override
-        public String get(CharSequence name) {
-            return getHeader(name.toString());
-        }
-
-        @Override
-        public Set<String> names() {
-            return headers.keySet();
-        }
-
-        @Override
-        public Collection<List<String>> values() {
-            return headers.values();
-        }
-
-        @Override
-        public <T> Optional<T> get(CharSequence name, ArgumentConversionContext<T> conversionContext) {
-            return ConversionService.SHARED
-                .convert(name.toString(), conversionContext);
-        }
     }
 }
